# docs/apidocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SocialPlus Backend API",
    "description": "API a vývojové endpointy pro aplikaci SocialPlus",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
...
  },
  "paths": {
...
    "/api/scheduler/upcoming": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "description": "Získání nadcházejících naplánovaných příspěvků",
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 168,
              "default": 24
            },
            "in": "query",
            "name": "hours",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "postId": {
                        "type": "number"
                      },
                      "networkId": {
                        "type": "number"
                      },
                      "contentId": {
                        "type": "number"
                      },
                      "content": {
                        "type": "string"
                      },
                      "attachments": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "scheduledDate": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "networkType": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://0.0.0.0:8080",
      "description": "HTTP server"
    }
  ],
...
}